requirements.txt
mysql-connector-python==8.0.23
python-dotenv==0.19.2
datetime==4.3
